* arronax/desktopfile.py    - save(): test existence before renaming to avoid error msg
* arronax/editor.py
   - show window early
   - add locale dir to gettext.bindtextdomain() call
   - read_desktop_file(): show msg on error
* arronax/parsecli.py
   - don't use enum.auto() for compatibility with Python 3.4

+++ arronax/desktopfile.py
@@ -353,7 +353,8 @@ class DesktopFile(object):
         content = self.keyfile.to_data()[0]
 
         try:
-            os.rename(path, '%s~' % path)
+            if os.path.exists(path):
+                os.rename(path, '%s~' % path)
         except Exception as e:
             logging.error('os.rename: p:{} e:{}'.format(path, e))
         try:
